extract activity_actor subclasses to a seperate header to reduce inclusion overhead and fanout (#45967)
