+++ src/avatar_action.cpp
@@ -14,6 +14,7 @@
 
 #include "action.h"
 #include "activity_actor.h"
+#include "activity_actor_definitions.h"
 #include "avatar.h"
 #include "bodypart.h"
 #include "cached_options.h"
