+++ src/handle_liquid.cpp
@@ -12,6 +12,7 @@
 
 #include "action.h"
 #include "activity_actor.h"
+#include "activity_actor_definitions.h"
 #include "activity_type.h"
 #include "cached_options.h"
 #include "cata_utility.h"
