+++ src/iexamine.cpp
@@ -12,6 +12,7 @@
 #include <utility>
 
 #include "activity_actor.h"
+#include "activity_actor_definitions.h"
 #include "activity_type.h"
 #include "ammo.h"
 #include "avatar.h"
