+++ src/iuse_actor.cpp
@@ -12,6 +12,7 @@
 #include <sstream>
 
 #include "action.h"
+#include "activity_actor_definitions.h"
 #include "activity_handlers.h"
 #include "activity_type.h"
 #include "assign.h"
