+++ src/monexamine.cpp
@@ -7,6 +7,7 @@
 #include <utility>
 #include <vector>
 
+#include "activity_actor_definitions.h"
 #include "avatar.h"
 #include "calendar.h"
 #include "cata_utility.h"
