+++ src/pickup.cpp
@@ -11,6 +11,7 @@
 #include <vector>
 
 #include "activity_actor.h"
+#include "activity_actor_definitions.h"
 #include "auto_pickup.h"
 #include "cata_utility.h"
 #include "catacharset.h"
