+++ src/ranged.cpp
@@ -13,6 +13,7 @@
 #include <vector>
 
 #include "activity_actor.h"
+#include "activity_actor_definitions.h"
 #include "avatar.h"
 #include "ballistics.h"
 #include "bodypart.h"
